requirements.txt
fastapi
pytest
snowflake-connector-python